+++ client_test.go
@@ -26,6 +26,7 @@ func (t *closureTransport) RoundTrip(req *http.Request) (*http.Response, error) 
 func TestUserAgentHeader(t *testing.T) {
 	var userAgent string
 	client := taamcloud.NewClient(
+		option.WithBearerToken("My Bearer Token"),
 		option.WithHTTPClient(&http.Client{
 			Transport: &closureTransport{
 				fn: func(req *http.Request) (*http.Response, error) {
@@ -38,7 +39,7 @@ func TestUserAgentHeader(t *testing.T) {
 		}),
 	)
 	client.Embeddings.New(context.Background(), taamcloud.EmbeddingNewParams{
-		Input: taamcloud.F([]string{"string"}),
+		Input: taamcloud.F([]string{"Generate vector representations of this text"}),
 		Model: taamcloud.F("jina-embeddings-v3"),
 	})
 	if userAgent != fmt.Sprintf("TaamCloud/Go %s", internal.PackageVersion) {
@@ -49,6 +50,7 @@ func TestUserAgentHeader(t *testing.T) {
 func TestRetryAfter(t *testing.T) {
 	retryCountHeaders := make([]string, 0)
 	client := taamcloud.NewClient(
+		option.WithBearerToken("My Bearer Token"),
 		option.WithHTTPClient(&http.Client{
 			Transport: &closureTransport{
 				fn: func(req *http.Request) (*http.Response, error) {
@@ -64,7 +66,7 @@ func TestRetryAfter(t *testing.T) {
 		}),
 	)
 	_, err := client.Embeddings.New(context.Background(), taamcloud.EmbeddingNewParams{
-		Input: taamcloud.F([]string{"string"}),
+		Input: taamcloud.F([]string{"Generate vector representations of this text"}),
 		Model: taamcloud.F("jina-embeddings-v3"),
 	})
 	if err == nil {
@@ -85,6 +87,7 @@ func TestRetryAfter(t *testing.T) {
 func TestDeleteRetryCountHeader(t *testing.T) {
 	retryCountHeaders := make([]string, 0)
 	client := taamcloud.NewClient(
+		option.WithBearerToken("My Bearer Token"),
 		option.WithHTTPClient(&http.Client{
 			Transport: &closureTransport{
 				fn: func(req *http.Request) (*http.Response, error) {
@@ -101,7 +104,7 @@ func TestDeleteRetryCountHeader(t *testing.T) {
 		option.WithHeaderDel("X-Stainless-Retry-Count"),
 	)
 	_, err := client.Embeddings.New(context.Background(), taamcloud.EmbeddingNewParams{
-		Input: taamcloud.F([]string{"string"}),
+		Input: taamcloud.F([]string{"Generate vector representations of this text"}),
 		Model: taamcloud.F("jina-embeddings-v3"),
 	})
 	if err == nil {
@@ -117,6 +120,7 @@ func TestDeleteRetryCountHeader(t *testing.T) {
 func TestOverwriteRetryCountHeader(t *testing.T) {
 	retryCountHeaders := make([]string, 0)
 	client := taamcloud.NewClient(
+		option.WithBearerToken("My Bearer Token"),
 		option.WithHTTPClient(&http.Client{
 			Transport: &closureTransport{
 				fn: func(req *http.Request) (*http.Response, error) {
@@ -133,7 +137,7 @@ func TestOverwriteRetryCountHeader(t *testing.T) {
 		option.WithHeader("X-Stainless-Retry-Count", "42"),
 	)
 	_, err := client.Embeddings.New(context.Background(), taamcloud.EmbeddingNewParams{
-		Input: taamcloud.F([]string{"string"}),
+		Input: taamcloud.F([]string{"Generate vector representations of this text"}),
 		Model: taamcloud.F("jina-embeddings-v3"),
 	})
 	if err == nil {
@@ -149,6 +153,7 @@ func TestOverwriteRetryCountHeader(t *testing.T) {
 func TestRetryAfterMs(t *testing.T) {
 	attempts := 0
 	client := taamcloud.NewClient(
+		option.WithBearerToken("My Bearer Token"),
 		option.WithHTTPClient(&http.Client{
 			Transport: &closureTransport{
 				fn: func(req *http.Request) (*http.Response, error) {
@@ -164,7 +169,7 @@ func TestRetryAfterMs(t *testing.T) {
 		}),
 	)
 	_, err := client.Embeddings.New(context.Background(), taamcloud.EmbeddingNewParams{
-		Input: taamcloud.F([]string{"string"}),
+		Input: taamcloud.F([]string{"Generate vector representations of this text"}),
 		Model: taamcloud.F("jina-embeddings-v3"),
 	})
 	if err == nil {
@@ -177,6 +182,7 @@ func TestRetryAfterMs(t *testing.T) {
 
 func TestContextCancel(t *testing.T) {
 	client := taamcloud.NewClient(
+		option.WithBearerToken("My Bearer Token"),
 		option.WithHTTPClient(&http.Client{
 			Transport: &closureTransport{
 				fn: func(req *http.Request) (*http.Response, error) {
@@ -189,7 +195,7 @@ func TestContextCancel(t *testing.T) {
 	cancelCtx, cancel := context.WithCancel(context.Background())
 	cancel()
 	_, err := client.Embeddings.New(cancelCtx, taamcloud.EmbeddingNewParams{
-		Input: taamcloud.F([]string{"string"}),
+		Input: taamcloud.F([]string{"Generate vector representations of this text"}),
 		Model: taamcloud.F("jina-embeddings-v3"),
 	})
 	if err == nil {
@@ -199,6 +205,7 @@ func TestContextCancel(t *testing.T) {
 
 func TestContextCancelDelay(t *testing.T) {
 	client := taamcloud.NewClient(
+		option.WithBearerToken("My Bearer Token"),
 		option.WithHTTPClient(&http.Client{
 			Transport: &closureTransport{
 				fn: func(req *http.Request) (*http.Response, error) {
@@ -211,7 +218,7 @@ func TestContextCancelDelay(t *testing.T) {
 	cancelCtx, cancel := context.WithTimeout(context.Background(), 2*time.Millisecond)
 	defer cancel()
 	_, err := client.Embeddings.New(cancelCtx, taamcloud.EmbeddingNewParams{
-		Input: taamcloud.F([]string{"string"}),
+		Input: taamcloud.F([]string{"Generate vector representations of this text"}),
 		Model: taamcloud.F("jina-embeddings-v3"),
 	})
 	if err == nil {
@@ -229,6 +236,7 @@ func TestContextDeadline(t *testing.T) {
 
 	go func() {
 		client := taamcloud.NewClient(
+			option.WithBearerToken("My Bearer Token"),
 			option.WithHTTPClient(&http.Client{
 				Transport: &closureTransport{
 					fn: func(req *http.Request) (*http.Response, error) {
@@ -239,7 +247,7 @@ func TestContextDeadline(t *testing.T) {
 			}),
 		)
 		_, err := client.Embeddings.New(deadlineCtx, taamcloud.EmbeddingNewParams{
-			Input: taamcloud.F([]string{"string"}),
+			Input: taamcloud.F([]string{"Generate vector representations of this text"}),
 			Model: taamcloud.F("jina-embeddings-v3"),
 		})
 		if err == nil {